Moved loading and parsing savefiles to SaveManager class
