+++ Major Production/Assets/_Scripts/Save System/SaveManager.cs.meta
@@ -1,0 +1,13 @@
+fileFormatVersion: 2
+guid: 4453701e6b9deca4990d6ae039211325
+timeCreated: 1536632304
+licenseType: Free
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
